+++ frontend/src/views/WasteActionsPage/WasteActionPage.vue
@@ -32,7 +32,6 @@
           <p v-html="wasteAction.description" class="mt-9"></p>
         </v-col>
         <v-col v-if="loggedUser" cols="12" class="d-flex flex-column align-start mt-8" sm="2">
-          <!-- Implement action buttons -->
           <p class="mb-2">Mis en place</p>
           <DsfrTagGroup
             v-if="canteensDoneAction && canteensDoneAction.length"
@@ -55,58 +54,30 @@
       <v-row class="mt-9">
         <BackLink :to="backLink" text="Retour" :primary="true" />
       </v-row>
-
-      <v-dialog v-model="actionDialog">
-        <v-card class="pa-6">
-          <div class="mt-n6 mx-n6 mb-4 pa-4 d-flex" style="background-color: #F5F5F5">
-            <v-spacer></v-spacer>
-            <v-btn outlined color="primary" @click="actionDialog = false">
-              Fermer
-            </v-btn>
-          </div>
-
-          <h2 id="modal-title" class="mb-3" tabindex="-1">Mis en place ?</h2>
-
-          <v-form ref="form" v-model="actionFormIsValid" @submit.prevent>
-            <DsfrAutocomplete
-              v-model="chosenCanteenIds"
-              :items="userCanteens"
-              clearable
-              multiple
-              hide-details
-              id="select-canteen"
-              placeholder="Toutes mes cantines"
-              class="mt-1"
-              no-data-text="Pas de résultats"
-              item-text="name"
-              item-value="id"
-            />
-            <v-row class="mt-2 pa-4">
-              <v-spacer></v-spacer>
-              <v-btn x-large color="primary" @click="saveActionChanges">Valider</v-btn>
-            </v-row>
-          </v-form>
-        </v-card>
-      </v-dialog>
+      <ResourceActionDialog
+        v-model="actionDialog"
+        :resourceId="id"
+        :userCanteens="userCanteens"
+        :actionsDone="actionsDone"
+        @close="closeActionDialog($event)"
+      />
     </div>
   </div>
 </template>
 <script>
-import DsfrAutocomplete from "@/components/DsfrAutocomplete"
 import BreadcrumbsNav from "@/components/BreadcrumbsNav.vue"
 import BackLink from "@/components/BackLink"
 import DsfrTagGroup from "@/components/DsfrTagGroup"
 import DsfrTag from "@/components/DsfrTag"
+import ResourceActionDialog from "@/components/ResourceActionDialog"
 import Constants from "@/constants"
 
 export default {
-  components: { DsfrAutocomplete, BreadcrumbsNav, BackLink, DsfrTagGroup, DsfrTag },
+  components: { BreadcrumbsNav, BackLink, DsfrTagGroup, DsfrTag, ResourceActionDialog },
   data() {
     return {
       wasteAction: null,
       actionDialog: false,
-      actionFormIsValid: true,
-      chosenCanteenIds: [],
       backLink: { name: "WasteActionsHome" },
     }
   },
@@ -139,43 +110,12 @@ export default {
           })
         })
     },
-    createOrUpdateResourceAction(canteenId, is_done = true) {
-      return this.$store
-        .dispatch("createOrUpdateResourceAction", {
-          resourceId: this.id,
-          payload: { canteenId, is_done },
-        })
-        .catch((e) => this.$store.dispatch("notifyServerError", e))
-    },
     showActionDialog() {
-      // Get the canteens that have already done the action
-      this.chosenCanteenIds = this.userCanteens
-        .filter((canteen) => this.actionsDone.find((actionCanteen) => actionCanteen.id === canteen.id))
-        .map((canteen) => canteen.id)
       this.actionDialog = true
     },
-    saveActionChanges() {
-      const actionChanges = []
-      // Compare the chosen canteens with the actions done (new & removed)
-      this.chosenCanteenIds.forEach((canteenId) => {
-        if (!this.actionsDone.find((actionCanteen) => actionCanteen.id === canteenId)) {
-          actionChanges.push(this.createOrUpdateResourceAction(canteenId, true))
-        }
-      })
-      this.actionsDone.forEach((actionCanteen) => {
-        if (!this.chosenCanteenIds.includes(actionCanteen.id)) {
-          actionChanges.push(this.createOrUpdateResourceAction(actionCanteen.id, false))
-        }
-      })
-      // close dialog and refresh the wasteAction if needed
-      if (actionChanges.length) {
-        Promise.all(actionChanges).then(() => {
-          this.fetchWasteAction()
-          this.actionDialog = false
-        })
-      } else {
-        this.actionDialog = false
-      }
+    closeActionDialog(refresh) {
+      if (refresh) this.fetchWasteAction()
+      this.actionDialog = false
     },
   },
   computed: {
